WIP update
Items in manual_hex is an attempt at parsing LMC 8.4 for possible values that aren't specified in Rivov API.

libValuesAPI now echoes output file as soon as it's opened, so we can pull it for analysis or mark it as invalid as quickly as possible.

ExposureGain array added to libparams.json - was in API as hexstring and needed to be parsed into individual floats.